requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
jinja2
passlib[bcrypt]
itsdangerous
jwt
pydantic[email]
python-multipart
yagmail
pandas

